#88 Renamed bezier.h to bezier_curve.h
